ValidationUserExist

+++ src/Modules/User/UserService.ts
@@ -34,16 +34,30 @@ export default class userService implements IUserService{
 
     async findById(id: string): Promise<User | Error> {
         const user = await this.userRepository.findById(id)
+        
+        if(!user){
+            return new Error("User does not exists!")
+        }
+
         return user
     }
 
     async delete(id: string): Promise<void | Error> {
+
+        if(!(await this.userRepository.findById(id))){
+            return new Error("User does not exists!")
+        }
+
         await this.userRepository.delete(id)
     }
 
     async update(id: string, name?: string, CPF?: string, email?: string): Promise<object | Error> {
         
         const user = await this.userRepository.update(id,name,CPF,email)
+
+        if(!(await this.userRepository.findById(id))){
+            return new Error("User does not exists!")
+        }
 
         return user
